+++ src/form/common/default-validation-conditionals.ts
@@ -112,6 +112,20 @@ export const hideIfInformantMotherOrFather = [
 export const isInformantSpouse =
   '!values.informantType || values.informantType==="SPOUSE"'
 
+export const hideIfDeceasedAddressNotAvailable = [
+  {
+    action: 'hide',
+    expression: '!(draftData && draftData.deceased?.countryPrimaryDeceased)'
+  }
+]
+
+export const hideIfMotherAddressNotAvailable = [
+  {
+    action: 'hide',
+    expression: '!(draftData && draftData.mother?.countryPrimaryMother)'
+  }
+]
+
 export const hideIfInformantSpouse = [
   {
     action: 'hide',
@@ -303,7 +317,7 @@ export function getNationalIDValidators(configCase: string): Validator[] {
       },
       {
         operation: 'duplicateIDNumber',
-        parameters: ['mother.iD']
+        parameters: ['mother.motherNationalId']
       }
     ]
   } else if (configCase === 'mother') {
@@ -314,7 +328,7 @@ export function getNationalIDValidators(configCase: string): Validator[] {
       },
       {
         operation: 'duplicateIDNumber',
-        parameters: ['father.iD']
+        parameters: ['father.fatherNationalId']
       }
     ]
   } else if (configCase === 'deceased') {
@@ -325,7 +339,7 @@ export function getNationalIDValidators(configCase: string): Validator[] {
       },
       {
         operation: 'duplicateIDNumber',
-        parameters: ['informant.informantID']
+        parameters: ['informant.informantNationalId']
       }
     ]
   } else if (configCase === 'groom') {
@@ -336,7 +350,7 @@ export function getNationalIDValidators(configCase: string): Validator[] {
       },
       {
         operation: 'duplicateIDNumber',
-        parameters: ['bride.iD']
+        parameters: ['bride.brideNationalId']
       }
     ]
   } else if (configCase === 'bride') {
@@ -347,7 +361,7 @@ export function getNationalIDValidators(configCase: string): Validator[] {
       },
       {
         operation: 'duplicateIDNumber',
-        parameters: ['groom.iD']
+        parameters: ['groom.groomNationalId']
       }
     ]
   } else {
@@ -359,23 +373,23 @@ export function getNationalIDValidators(configCase: string): Validator[] {
       },
       {
         operation: 'duplicateIDNumber',
-        parameters: ['deceased.deceasedID']
+        parameters: ['deceased.deceasedNationalId']
       },
       {
         operation: 'duplicateIDNumber',
-        parameters: ['mother.iD']
+        parameters: ['mother.motherNationalId']
       },
       {
         operation: 'duplicateIDNumber',
-        parameters: ['father.iD']
+        parameters: ['father.fatherNationalId']
       },
       {
         operation: 'duplicateIDNumber',
-        parameters: ['groom.iD']
+        parameters: ['groom.groomNationalId']
       },
       {
         operation: 'duplicateIDNumber',
-        parameters: ['bride.iD']
+        parameters: ['bride.brideNationalId']
       }
     ]
   }